Add GNU General Public License v3
This commit adds the GNU General Public License version 3 to the project. This license is a free, copyleft license for software and other kinds of works that guarantees users' freedom to share and change all versions of a program. The license includes terms and conditions for copying, distribution, and modification of the work.